Remove getter file Added get_number_map_db in config
